Update calibration processing to use EV results for single-target polar plots
Previously the polar plots used data from the EK80 .xml files, which don't use the calibration results used for the survey, which is now typically done through Echoview. This script has been updated so that the it now use calibrated single-target detections exported from Echoview to make the polar plots.  This requires pointing to additional data files, which are now specified in the settings file.